requirements.txt
six
argparse
babel
iso8601>=0.1.9
pbr>=0.11,<2.0
PrettyTable>=0.7,<0.8
python-keystoneclient>=0.9.0
PyYAML>=3.1.0
requests>=1.1
